### 2022-04-01 #4295 @kyleniemeyer
Hi @warrickball, looks like there is an issue with part of the header in your paper. Could you fix
```
authors:
  - name: Warrick H. Ball
    orcid: 0000-0002-4773-1017
affiliations:
 - name: School of Physics & Astronomy, University of Birmingham, United Kingdom
```
to
```
authors:
  - name: Warrick H. Ball
    orcid: 0000-0002-4773-1017
    affiliation: 1
affiliations:
 - name: School of Physics & Astronomy, University of Birmingham, United Kingdom
   index: 1
```

and also take a look at some of those reference issues above?
