### 2022-04-01 24937d3
joss-paper: fix affiliation, remove unused references, point ADS URLs to main server
