+++ joss-paper/paper.bib
@@ -16,7 +16,7 @@
    volume = 272,
     pages = {1286-1292},
       doi = {10.1126/science.272.5266.1286},
-   adsurl = {http://esoads.eso.org/abs/1996Sci...272.1286C},
+   adsurl = {https://ui.adsabs.harvard.edu/abs/1996Sci...272.1286C},
   adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
 
@@ -30,7 +30,7 @@ booktitle = {Asteroseismology},
      isbn = {978-1-4020-5178-4},
 publisher = {Springer, Berlin},
      year = 2010,
-   adsurl = {http://esoads.eso.org/abs/2010aste.book.....A},
+   adsurl = {https://ui.adsabs.harvard.edu/abs/2010aste.book.....A},
   adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
 
@@ -49,230 +49,3 @@ publisher = {Springer, Berlin},
        adsurl = {https://ui.adsabs.harvard.edu/abs/2000MNRAS.316..165M},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
-
-
-@ARTICLE{adipls,
-   author = {{Christensen-Dalsgaard}, J.},
-    title = "{ADIPLS--the Aarhus adiabatic oscillation package}",
-  journal = {\apss},
-archivePrefix = "arXiv",
-   eprint = {0710.3106},
- keywords = {Stars: oscillations, Numerical methods, Asteroseismology},
-     year = 2008,
-    month = aug,
-   volume = 316,
-    pages = {113-120},
-      doi = {10.1007/s10509-007-9689-z},
-   adsurl = {http://adsabs.harvard.edu/abs/2008Ap%26SS.316..113C},
-  adsnote = {Provided by the SAO/NASA Astrophysics Data System}
-}
-
-
-@ARTICLE{gyre1,
-   author = {{Townsend}, R.~H.~D. and {Teitler}, S.~A.},
-    title = "{GYRE: an open-source stellar oscillation code based on a new Magnus Multiple Shooting scheme}",
-  journal = {\mnras},
-archivePrefix = "arXiv",
-   eprint = {1308.2965},
- primaryClass = "astro-ph.SR",
- keywords = {methods: numerical, stars: evolution, stars: interiors, stars: oscillations, stars: variables: general},
-     year = 2013,
-    month = nov,
-   volume = 435,
-    pages = {3406-3418},
-      doi = {10.1093/mnras/stt1533},
-   adsurl = {http://ukads.nottingham.ac.uk/abs/2013MNRAS.435.3406T},
-  adsnote = {Provided by the SAO/NASA Astrophysics Data System}
-}
-
-
-@ARTICLE{gyre2,
-   author = {{Townsend}, R.~H.~D. and {Goldstein}, J. and {Zweibel}, E.~G.
-	},
-    title = "{Angular momentum transport by heat-driven g-modes in slowly pulsating B stars}",
-  journal = {\mnras},
-archivePrefix = "arXiv",
-   eprint = {1712.02420},
- primaryClass = "astro-ph.SR",
- keywords = {asteroseismology, stars: evolution, stars: interiors, stars: massive, stars: oscillations, stars: rotation},
-     year = 2018,
-    month = mar,
-   volume = 475,
-    pages = {879-893},
-      doi = {10.1093/mnras/stx3142},
-   adsurl = {http://ukads.nottingham.ac.uk/abs/2018MNRAS.475..879T},
-  adsnote = {Provided by the SAO/NASA Astrophysics Data System}
-}
-
-
-@ARTICLE{eggleton1971,
-   author = {{Eggleton}, P.~P.},
-    title = "{The evolution of low mass stars}",
[…]
