### 2022-04-01 5b23a4d
add DOI for Asteroseismology book (from HTML source for SpringerLink page)
