+++ joss-paper/paper.bib
@@ -30,6 +30,7 @@ booktitle = {Asteroseismology},
      isbn = {978-1-4020-5178-4},
 publisher = {Springer, Berlin},
      year = 2010,
+      doi = {10.1007/978-1-4020-5803-5},
    adsurl = {https://ui.adsabs.harvard.edu/abs/2010aste.book.....A},
   adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
