### 2022-04-01 #4295 @warrickball
> Hi @warrickball, looks like there is an issue with part of the header in your paper.

Fixed, thanks!

> and also take a look at some of those reference issues above?

I fixed most of this by removing the many references that were in discarded drafts and looked (quite thoroughly...) to find the DOI for the Asteroseismology textbook. (The suggested DOI was for a different book.)
