### 2022-04-04 #4295 @christinahedges
Thanks @kyleniemeyer I'm happy to edit this submission!

@warrickball thank you for your submission to JOSS. I'm going to get to work checking through the submission, and looking for appropriate reviewers. If you know there are people in your field who are not [conflicted with you](https://joss.readthedocs.io/en/latest/review_checklist.html?highlight=conflict#conflict-of-interest) please let me know and I can use that as a starting point to find reviewers.

A quick point with your submission paper, if there are any other tools that can perform similar functionality, you should mention them in the submission paper. Since tosmo is an interface for other algorithms, you may want to call them out by name?

I'll start looking for reviewers!
