### 2022-04-05 #4295 @warrickball
> If you know there are people in your field who are not [conflicted with you](https://joss.readthedocs.io/en/latest/review_checklist.html?highlight=conflict#conflict-of-interest) please let me know and I can use that as a starting point to find reviewers.

I'll give it some thought. Is it okay to email you some suggestions (once I've thought of a helpful number)? I'm not sure how comfortable they might be having their names attached here (in public) without their consent.

> A quick point with your submission paper, if there are any other tools that can perform similar functionality, you should mention them in the submission paper. Since tosmo is an interface for other algorithms, you may want to call them out by name?

I'll do that, thanks!
