### 2022-04-19 #4295 @christinahedges
Hi @warrickball, I've added another reviewer to the list. I know we spoke briefly about possible conflicts of interest with reviewers. Asteroseismology as a field is very collaborative (excellent!) but that can make it hard to find completely unconflicted reviewers. @danhey and @astrobel are authors on papers that you are also an author on in the last two years, but I have 1) spoken with the editorial team and 2) spoken with @danhey and @astrobel and we feel that, because your contributions on these works were well separated, you have only very minor conflicts, and we are happy to continue with them as reviewers.

@warrickball, to verify, if you feel that there are significant enough conflicts of interest that we should find new reviewers, please let me know here or on email, otherwise I will kick off the review with these two asteroseismology and coding expert reviewers!

@danhey and @astrobel thank you for accepting and offering your time!
