### 2022-04-25 #4343 @editorialbot
[REVIEW]: tomso: TOols for Models of Stars and their Oscillations
**Submitting author:** @warrickball (<a href="http://orcid.org/0000-0002-4773-1017">Warrick Ball</a>)
**Repository:** https://github.com/warrickball/tomso
**Branch with paper.md** (empty if default branch): joss-paper
**Version:** v0.2.0
