### 2022-04-26 a879bf9
delete duplicate Fig.; name supported programs, with references
