+++ joss-paper/paper.bib
@@ -50,3 +50,75 @@ publisher = {Springer, Berlin},
        adsurl = {https://ui.adsabs.harvard.edu/abs/2000MNRAS.316..165M},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
+
+
+@ARTICLE{adipls,
+   author = {{Christensen-Dalsgaard}, J.},
+    title = "{ADIPLS--the Aarhus adiabatic oscillation package}",
+  journal = {\apss},
+archivePrefix = "arXiv",
+   eprint = {0710.3106},
+ keywords = {Stars: oscillations, Numerical methods, Asteroseismology},
+     year = 2008,
+    month = aug,
+   volume = 316,
+    pages = {113-120},
+      doi = {10.1007/s10509-007-9689-z},
+   adsurl = {http://adsabs.harvard.edu/abs/2008Ap%26SS.316..113C},
+  adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
+
+@ARTICLE{gyre,
+   author = {{Townsend}, R.~H.~D. and {Teitler}, S.~A.},
+    title = "{GYRE: an open-source stellar oscillation code based on a new Magnus Multiple Shooting scheme}",
+  journal = {\mnras},
+archivePrefix = "arXiv",
+   eprint = {1308.2965},
+ primaryClass = "astro-ph.SR",
+ keywords = {methods: numerical, stars: evolution, stars: interiors, stars: oscillations, stars: variables: general},
+     year = 2013,
+    month = nov,
+   volume = 435,
+    pages = {3406-3418},
+      doi = {10.1093/mnras/stt1533},
+   adsurl = {http://ukads.nottingham.ac.uk/abs/2013MNRAS.435.3406T},
+  adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
+
+@ARTICLE{mesa,
+   author = {{Paxton}, B. and {Bildsten}, L. and {Dotter}, A. and {Herwig}, F. and 
+	{Lesaffre}, P. and {Timmes}, F.},
+    title = "{Modules for Experiments in Stellar Astrophysics (MESA)}",
+  journal = {\apjs},
+archivePrefix = "arXiv",
+   eprint = {1009.1622},
+ primaryClass = "astro-ph.SR",
+ keywords = {methods: numerical, stars: evolution, stars: general},
+     year = 2011,
+    month = jan,
+   volume = 192,
+      eid = {3},
+    pages = {3},
+      doi = {10.1088/0067-0049/192/1/3},
+   adsurl = {http://esoads.eso.org/abs/2011ApJS..192....3P},
+  adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
+
+@MISC{stars,
+       author = {{Eggleton}, P.~P. and {Tout}, Christopher and {Pols}, Onno and {Izzard}, Rob and {Eldridge}, John and {Lesaffre}, Pierre and {Stancliffe}, Richard and {Church}, Ross and {Lau}, Herbert},
+        title = "{STARS: A Stellar Evolution Code}",
+     keywords = {Software},
+ howpublished = {Astrophysics Source Code Library, record ascl:1107.008},
+         year = 2011,
+        month = jul,
+          eid = {ascl:1107.008},
+        pages = {ascl:1107.008},
+archivePrefix = {ascl},
+       eprint = {1107.008},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2011ascl.soft07008E},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
