+++ joss-paper/paper.md
@@ -36,7 +36,11 @@ Some are fixed-width plain text files with several blocks; some are
 Fortran binaries; and few can easily be read with standard routines
 for loading data.  Some programs require data to be prepared in a
 specific format and provide tools to do so but these tools are
-incomplete and difficult to extend.
+incomplete and difficult to extend.  For example, the Aarhus adiabatic
+oscillation package [ADIPLS, @adipls] includes tools to convert
+stellar models of one particular widely-used format (FGONG) into its
+internal Fortran binary format but not to convert them back after
+they've been manipulated by other bundled tools.
 
 `tomso` is a set of Python modules that provides a high-level
 interface to several formats and enables or simplifies several common
@@ -62,11 +66,15 @@ Such a figure is shown for a standard solar model, Model S [@modelS],
 in the textbook by @acdk2010 [their Fig. 7.30, which is in essence the
 same as Fig. 1 in @monteiro2000] but requires some manipulation of the
 source data because neither $c_\mathrm{s}$ nor $\tau$ are part of the
-data format.  Fig. \autoref{fig:dc_dtau} shows the same data but with
+data format.  \autoref{fig:dc_dtau} shows the same data but with
 `tomso` this is naturally expressed in two functional lines of Python code: one
 to read the data file and one to plot the relevant data using the
 high-level properties `cs` and `tau`.
 
+`tomso` currently includes interfaces for ADIPLS, the FGONG file format,
+output from GYRE [@gyre], input and output from the Modules for
+Experiments in Stellar Astrophysics [MESA, @mesa], and output from
+the Cambridge stellar evolution code [STARS, @stars].
 `tomso`'s interfaces also aim to be easily extensible, so that in the
 future it can not only support more current codes and file formats,
 but also those that have yet to be developed.
