### 2022-04-26 #4343 @warrickball
I've explicitly named and cited (as far as possible) the currently supported programs and formats in the paper, as @christinahedges [suggested in the pre-review](https://github.com/openjournals/joss-reviews/issues/4295#issuecomment-1088097878).
