### 2022-05-05 #4343 @danhey
### Intro Comments:

- Thanks for inviting me to review `tomso`. This is a nice piece of software and a solid JOSS submission that will bridge the gap between the different stellar evolution code models. I've been playing around with the code for a while and it all runs nicely. My only issues relate to some lacking documentation and comparisons to other similar codes. I discuss these below.

### Installation and docs

- Both pip install and from source works fine on a fresh Python 3.7 environment.
- I am a bit alarmed at the line in the introduction that says the code is currently very unstable and the API will change drastically. Has this already happened?
- There are community guidelines on the repository but not in the documentation. I don't think this is hugely important, but it's nice to have a short section after the installation. There should also be a small example on running the unit tests (which isn't always obvious!).
- It would be nice to have some examples for the STARS models -- its inclusion in the Code Outputs page currently seems like an afterthought.
- On that note, is STARS included in the Stellar Models page? I must admit I am unfamiliar with these models, but it looks like `tomso` can not convert between them?

### Unit tests:

- The unit tests are nicely done and comprehensive. This is a great example of using GitHub workflows to automate testing -- nice work! This is only a suggestion, but it would be nice to have a badge on the repo showing the test status (i.e., guide [here](https://docs.github.com/en/actions/monitoring-and-troubleshooting-workflows/adding-a-workflow-status-badge))

### Repository:

- All good.

### Examples:

- I get the same output running the examples in the documentation.
- I do feel that the documentation is slightly lacking in some examples. It would be nice to have a few more examples for the command line interface and stellar models pages.

Manuscript:

- The manuscript is clear and concise.
- My only issue here is that no discussion is made of similar packages (i.e., [PyGyre](https://github.com/rhdtownsend/pygyre)) and how they differ. Obviously `tomso` fills a different role, but there is some overlap in capabilities.
- Some Googling around shows a few similar packages for reading in model data, such as [PyMesa](https://github.com/rjfarmer/pyMesa) and the [mesa_web](http://www.astro.wisc.edu/~townsend/static.php?ref=mesa-web-output) script.
