### 2022-05-06 22f8bd1
increase density of MESA test history so that example plot looks better
