### 2022-05-06 2fd8592
add basic support for command line plots of STARS plot files
