### 2022-05-06 d961f20
fix/expand documentation, particularly on script usage and STARS output
