### 2022-05-06 #4343 @warrickball
Thanks for the review @danhey! I've tried to address everything you raised about the repository itself but haven't yet gotten to the paper. Specifically,
* I removed the line about the unstable API. Yes, this change (making things more object-oriented) has now (mostly) happened.
* I copied the community guidelines to the docs' index page.
* I extended the description of the STARS output and added some examples to the Code Outputs page and the Command-line Interface page (which involved tweaking the script to actually support STARS' `plot` files).
* I added some badges to the front page (by shamelessly copying [emcee](https://emcee.readthedocs.io/)).

I'll work on the manuscript later.

> * On that note, is STARS included in the Stellar Models page? I must admit I am unfamiliar with these models, but it looks like tomso can not convert between them?

The user defines the contents of STARS' profiles when they run the code, so there's no guarantee that all the information for a seismic stellar model is available. That said, it should be possible to extend `tomso` to at least *try* to convert a STARS profile to an seismo-friendly format and report an error if it can't. I'm going to hide behind that being "beyond the scope of the present work" but make a note to give it a try if I get the chance.
