+++ joss-paper/paper.bib
@@ -122,3 +122,27 @@ archivePrefix = {ascl},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
 
+
+@software{pymesa,
+  author       = {Robert Farmer and
+                  Evan B. Bauer},
+  title        = {rjfarmer/pyMesa: Add support for 10398},
+  month        = mar,
+  year         = 2018,
+  publisher    = {Zenodo},
+  version      = {v1.0.3},
+  doi          = {10.5281/zenodo.1205271},
+  url          = {https://doi.org/10.5281/zenodo.1205271}
+}
+
+
+@software{mesaplot,
+  author       = {Robert Farmer},
+  title        = {rjfarmer/mesaplot: Release: v2.0.3},
+  month        = dec,
+  year         = 2021,
+  publisher    = {Zenodo},
+  version      = {v2.0.3},
+  doi          = {10.5281/zenodo.5779536},
+  url          = {https://doi.org/10.5281/zenodo.5779536}
+}
