### 2022-05-06 #4343 @warrickball
@danhey I've fleshed out a bit of what the other programs do but let me know if this still isn't what you had in mind.
