### 2022-05-09 a873ae1
add basic support for command line plots of summaries from STARS out files
