+++ tomso/stars.py
@@ -106,7 +106,7 @@ summary_dtypes = [('n','int'), ('dt', 'float'),
 plot_dtypes = [('n','int'), ('age', 'float'), ('logR', 'float'),
                ('logTeff', 'float'), ('logL', 'float'), ('M', 'float'),
                ('M_H','float'), ('M_He', 'float'), ('L_H', 'float'),
-               ('L_He', 'float'), ('L_C', 'float'), ('mconv1', 'float'),
+               ('L_He', 'float'), ('L_C', 'float'), ('m_conv1', 'float'),
                ('m_conv2','float'), ('m_conv3','float'), ('m_conv4','float'),
                ('m_conv5','float'), ('m_conv6','float'), ('m_conv7','float'),
                ('m_conv8','float'), ('m_conv9','float'), ('m_conv10','float'),
