### 2022-05-11 #4343 @astrobel
Thanks @warrickball and JOSS for having me as a reviewer! This is a nifty software package, and everything works well for me! I've made some notes below on specific things I think can be clarified or improved. My only overall comment is that you should standardise the formatting of `tomso` across the paper and the docs. Right now, it's lowercase and verbatim in the paper and uppercase in the docs. I would suggest the lowercase version for uniformity, as this seems to be the choice on Github too.

### Documentation

In general, the documentation is very clear and complete. In particular, I think the API documentation is excellent. I only have several small notes:

* Please include the community guidelines in the docs as well as the README!
* I suggest including a sidebar link to `https://tomso.readthedocs.io/en/stable/index.html#installation`.
* I notice above that you plan to remove the alarming line about the unstable API, but I just wanted to echo @danhey's sentiment that this struck me as worrying. It would be worth noting something more along the lines of, keep an eye out for regular updates.
* I would suggest removing the links to "Source code" beneath examples that download .py files of the verbatim code blocks. This seems superfluous to me, and is non-uniformly present throughout the docs. I would also recommend removing the download links to various versions of the output plot, though I would say that's less of an issue.
* In the "Code outputs" section, I suggest including a brief example for handling STARS data, for the sake of completeness. It would also be good to link to the associated API page for each section.
* This is a nitpick, but in the docs page on the command-line interface, "which should be in your $PATH after installing TOMSO" can be reworded to something along the lines of "which will be in your $PATH after installing TOMSO. If it isn't, please contact me." It was in the $PATH for me!

### Software paper

The paper is a great summary of tomso's applications and the niche it fills in the field. However, there are some areas where I think you could tidy up the flow of information to better emphasise tomso's most noteworthy features, putting the important details at the forefront, as I've suggested below:

* I know that the summary is intended to be for a non-specialist audience, but I would nevertheless include something along the lines of "the study of resonant stellar pulsations is known as asteroseismology."
* It is worth emphasising in the summary that tomso is the _only_ software package that provides a unified interface for all the currently-used stellar oscillation code output and input files. You mention at the very end of the "Statement of need" section that tomso is extensible for future developments in the field. I also think this is a really impressive feature of the software that deserves brief mention in the summary.
* On that note, I find the sentence beginning in line 29 to be confusing and too vague. One possible rewording would be: "tomso is a set of Python modules that provides a compact interface to access data in multiple different formats. It also enables and simplifies several common tasks in analysing the stellar models and oscillations." (I changed the "or" to an "and" as a suggestion, because I believe enabling the tasks is also a form of simplification, but it does make sense to specify simplification separately!)
* The citation for the Aerts et al. textbook is distractingly long, and breaks up that sentence in a way that made it hard for me to parse. I suggest ending the sentence after the citation and beginning the next with "However, that figure requires some manipulation ..."
