### 2022-05-16 9fe2652
plots in docs: don't render links to source or alternative formats
