### 2022-05-16 846d67f
TOMSO -> tomso, and monotype as much as possible
