### 2022-05-16 #4343 @warrickball
@astrobel I've made a few changes to the paper that I hope address your points. In case it isn't coming across properly, `tomso` doesn't claim to support everything and extension would require writing code, rather than `tomso` being able to generate new interfaces (which would be awesome but not what it does). But yes, at least some of the extension would use existing infrastructure. e.g., parent classes from which the current suite of classes inherit.

The interfaces also aren't completely unified. They're basically unified for the file formats that are used as inputs for pulsation codes, because they have to contain certain types of information. But many of the output formats can contain whatever the user defines at runtime, which may or may not then be convertible.
