### 2022-05-17 #4343 @astrobel
Hi @warrickball! I've had a look through the paper and it looks great, and reads very clearly.

Given the issues with Sphinx, I would say no worries on that front. I'm certainly not going to make you mess around with the CSS for this!!

Thanks for all your work! Based on these updates, I'm happy to recommend the software for publication in JOSS.
