+++ HISTORY.txt
@@ -1,3 +1,4 @@
+v0.2.0,   2022-05-19 -- add basic support for command-line plots of STARS output
 v0.1.0,   2022-03-31 -- removed old procedural functions
           2021-06-15 -- added 'unique' option to `tomso plot --legend`
           2021-01-08 -- extend MESAAsteroSamples slicing to handle NumPy ints,
