### 2022-05-19 feebd5b
TOMSO -> tomso in README.md
