### 2022-05-19 070fbd3
TOMSO -> tomso in README.md
