### 2022-05-19 #4343 @warrickball
Thanks everyone! I've now been both sides of the JOSS review process and its always been great.

@christinahedges I added a few more lines to the paper to better distinguish `tomso` from other packages.

I think I've set up the Zenodo-GitHub integration correctly and released [v0.2.0](https://pypi.org/project/tomso/0.2.0/), which has DOI [10.5281/zenodo.6563220](https://doi.org/10.5281/zenodo.6563220). I apologise to @astrobel: I only commited changes TOMSO to `tomso` in the README *after* minting the package, so that'll only be fixed in the next version.
