### 2022-05-19 09246ba
remove old tomso-convert script, which has been replaced by convert subcommand of tomso script
