### 2022-05-23 #4343 @christinahedges
@warrickball, @astrobel and @danhey thanks so much participating in the JOSS review process! I'm going to recommend this submission for acceptance.

It seems like editorialbot is having a problem generating the pdf here. Maybe the EiC for this submission @kyleniemeyer can help us figure out what's going wrong with editorialbot? 😨
