### 2022-05-23 #4343 @kyleniemeyer
It looks like the issue is with the figure, which references a URL that does not resolve (https://tomso.readthedocs.io/en/stable/index-1.hires.png).

Can you instead put the png locally in the directory with the paper source file?
