### 2022-05-24 c0518e2
store figure in repo (with script to generate it)
