### 2022-05-24 #4343 @warrickball
Thanks @kyleniemeyer: I can see how this ended happening after I released the JOSS version of the package. The PNG (and short script to generate it) are now with the paper source. The draft PDF action [just passed](https://github.com/warrickball/tomso/actions/runs/2378357998).
