### 2022-05-24 acdc860
add missing NumPy import to example
