### 2022-05-24 e6016ef
Expand journal macros
