+++ joss-paper/paper.bib
@@ -39,7 +39,7 @@ publisher = {Springer, Berlin},
 @ARTICLE{monteiro2000,
        author = {{Monteiro}, M{\'a}rio J.~P.~F.~G. and {Christensen-Dalsgaard}, J{\o}rgen and {Thompson}, Michael J.},
         title = "{Seismic study of stellar convective regions: the base of the convective envelope in low-mass stars}",
-      journal = {\mnras},
+      journal = {MNRAS},
      keywords = {CONVECTION, STARS: EVOLUTION, STARS: INTERIORS, STARS: OSCILLATIONS},
          year = 2000,
         month = jul,
@@ -55,7 +55,7 @@ publisher = {Springer, Berlin},
 @ARTICLE{adipls,
    author = {{Christensen-Dalsgaard}, J.},
     title = "{ADIPLS--the Aarhus adiabatic oscillation package}",
-  journal = {\apss},
+  journal = {Ap&SS},
 archivePrefix = "arXiv",
    eprint = {0710.3106},
  keywords = {Stars: oscillations, Numerical methods, Asteroseismology},
@@ -72,7 +72,7 @@ archivePrefix = "arXiv",
 @ARTICLE{gyre,
    author = {{Townsend}, R.~H.~D. and {Teitler}, S.~A.},
     title = "{GYRE: an open-source stellar oscillation code based on a new Magnus Multiple Shooting scheme}",
-  journal = {\mnras},
+  journal = {MNRAS},
 archivePrefix = "arXiv",
    eprint = {1308.2965},
  primaryClass = "astro-ph.SR",
@@ -91,7 +91,7 @@ archivePrefix = "arXiv",
    author = {{Paxton}, B. and {Bildsten}, L. and {Dotter}, A. and {Herwig}, F. and 
 	{Lesaffre}, P. and {Timmes}, F.},
     title = "{Modules for Experiments in Stellar Astrophysics (MESA)}",
-  journal = {\apjs},
+  journal = {ApJS},
 archivePrefix = "arXiv",
    eprint = {1009.1622},
  primaryClass = "astro-ph.SR",
