### 2022-05-24 #4343 @warrickball
@kyleniemeyer Of course! I've replaced the macros with their standard abbreviations. i.e., `\mnras`→ MNRAS, `\apss` → Ap&SS and `\apjs` → ApJS.
